models/requirements.txt
opencv-python
onnxruntime
numpy
paho-mqtt
